Split scratchstack-database ops/iam/group up
