+++ scratchstack-database/src/ops/iam/group/add_user_to_group.rs
@@ -1,0 +1,107 @@
+//! AddUserToGroup database operation
+use {
+    crate::{
+        constants::iam::*,
+        ops::{
+            RequestExecutor,
+            iam::{validate_account_id, validate_group_name, validate_user_name},
+        },
+    },
+    indoc::indoc,
+    scratchstack_shapes_iam::{
+        error_meta::Error as IamError,
+        operation::AddUserToGroupInternalRequest,
+        types::error::{InternalFailure, NoSuchEntityException},
+    },
+    sqlx::{Row as _, postgres::PgTransaction, query},
+};
+
+impl RequestExecutor for AddUserToGroupInternalRequest {
+    type Response = ();
+    type Error = IamError;
+
+    async fn execute(&self, tx: &mut PgTransaction<'_>) -> Result<Self::Response, Self::Error> {
+        add_user_to_group(tx, &self.account_id, &self.group_name, &self.user_name).await
+    }
+}
+
+/// Add a user to a group in the database.
+pub async fn add_user_to_group(
+    tx: &mut PgTransaction<'_>,
+    account_id: &str,
+    group_name: &str,
+    user_name: &str,
+) -> Result<(), IamError> {
+    validate_account_id(account_id)?;
+    let account_id = match account_id {
+        AWS_ACCOUNT_ID => AWS_ACCOUNT_ID_NUMERIC,
+        account_id => account_id,
+    };
+    validate_group_name(group_name)?;
+    validate_user_name(user_name)?;
+
+    // Look up the group_id.
+    let group_id: String = match query(indoc! {"
+            SELECT group_id
+            FROM iam.groups
+            WHERE account_id = $1 AND group_name_lower = $2
+        "})
+    .bind(account_id)
+    .bind(group_name.to_lowercase())
+    .fetch_optional(tx.as_mut())
+    .await
+    {
+        Ok(Some(row)) => row.get(0),
+        Ok(None) => {
+            return Err(NoSuchEntityException::builder()
+                .message(format!("The group with name {group_name} cannot be found."))
+                .build()
+                .into());
+        }
+        Err(e) => {
+            log::error!("Failed to look up group in database: {e}");
+            return Err(InternalFailure::builder().message(MSG_INTERNAL_FAILURE).build().into());
+        }
+    };
+
+    // Look up the user_id.
+    let user_id: String = match query(indoc! {"
+            SELECT user_id
+            FROM iam.users
+            WHERE account_id = $1 AND user_name_lower = $2
+        "})
+    .bind(account_id)
+    .bind(user_name.to_lowercase())
+    .fetch_optional(tx.as_mut())
+    .await
+    {
+        Ok(Some(row)) => row.get(0),
+        Ok(None) => {
+            return Err(NoSuchEntityException::builder()
+                .message(format!("The user with name {user_name} cannot be found."))
+                .build()
+                .into());
+        }
+        Err(e) => {
+            log::error!("Failed to look up user in database: {e}");
+            return Err(InternalFailure::builder().message(MSG_INTERNAL_FAILURE).build().into());
+        }
+    };
+
+    // Insert the membership.
+    if let Err(e) = query(indoc! {"
+            INSERT INTO iam.group_memberships(group_id, user_id)
+            VALUES($1, $2)
+            ON CONFLICT (group_id, user_id) DO NOTHING
+        "})
+    .bind(&group_id)
+    .bind(&user_id)
+    .execute(tx.as_mut())
+    .await
+    {
+        log::error!("Failed to add user to group in database: {e}");
+        return Err(InternalFailure::builder().message(MSG_INTERNAL_FAILURE).build().into());
+    }
+
+    Ok(())
+}
